@@ -2,9 +2,10 @@
   "name": "credlify",
   "version": "1.3.0",
   "description": "A simple scaffolding helper for node projects",
+  "type": "module",
   "main": "credlify.js",
   "engines": {
-    "node": ">= 12.9.0"
+    "node": ">= 14.6.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
